Rename blob_client_options.hpp->blob_options.hpp
