+++ sdk/storage/inc/blobs/append_blob_client.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "blob_client_options.hpp"
+#include "blob_options.hpp"
 #include "blobs/blob_client.hpp"
 #include "common/storage_credential.hpp"
 #include "internal/protocol/blob_rest_client.hpp"
