+++ sdk/storage/inc/blobs/blob_container_client.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "blob_client_options.hpp"
+#include "blob_options.hpp"
 #include "blobs/blob_client.hpp"
 #include "common/storage_credential.hpp"
 #include "common/storage_url_builder.hpp"
